+++ e2e_tests/browser/pages/frontpage.page.ts
@@ -9,5 +9,5 @@ export const pageUrl = getPageUrl(RELATIVE_URL);
 
 // CSS selectors to find elements on a page. Used to construct Playwright `Locators`.
 export const locators = {
-  pageTitle: "[data-testid='lesson-title']",
+  pageTitle: "[data-testid='home-page-title']",
 };