+++ src/__tests__/__helpers__/renderWithProviders.tsx
@@ -7,8 +7,10 @@
  */
 import React, { FC, ReactElement } from "react";
 import { render, RenderOptions } from "@testing-library/react";
+import { MemoryRouterProvider } from "next-router-mock/MemoryRouterProvider";
 
 import { BookmarksProvider } from "../../hooks/useBookmarks";
+import { UserStyleContextProvider } from "../../context/UserStyleContext";
 import { SearchProvider } from "../../context/SearchContext";
 
 import MockedAuthProvider, {
@@ -27,9 +29,13 @@ const AllTheProviders: FC<ProviderProps> = ({
   return (
     <MockedAuthProvider {...authProviderProps}>
       <MockedApolloProvider>
-        <BookmarksProvider>
-          <SearchProvider>{children}</SearchProvider>
-        </BookmarksProvider>
+        <UserStyleContextProvider>
+          <MemoryRouterProvider>
+            <BookmarksProvider>
+              <SearchProvider>{children}</SearchProvider>
+            </BookmarksProvider>
+          </MemoryRouterProvider>
+        </UserStyleContextProvider>
       </MockedApolloProvider>
     </MockedAuthProvider>
   );