+++ src/pages/_app.tsx
@@ -8,9 +8,11 @@ import "../styles/reset.css";
 import "../styles/globals.css";
 import useApolloClient from "../browser-lib/graphql/useApolloClient";
 import useTheme from "../hooks/useTheme";
+import { UserStyleContextProvider } from "../context/UserStyleContext";
 import { SearchProvider } from "../context/SearchContext";
 import ErrorBoundary from "../components/ErrorBoundary";
 import { BookmarksProvider } from "../hooks/useBookmarks";
+import DefaultSeo from "../browser-lib/seo/DefaultSeo";
 
 const MyApp: FC<AppProps> = ({ Component, pageProps }) => {
   useTheme();
@@ -21,11 +23,14 @@ const MyApp: FC<AppProps> = ({ Component, pageProps }) => {
       <ErrorBoundary>
         <AuthProvider>
           <ApolloProvider client={apolloClient}>
-            <BookmarksProvider>
-              <SearchProvider>
-                <Component {...pageProps} />
-              </SearchProvider>
-            </BookmarksProvider>
+            <UserStyleContextProvider>
+              <BookmarksProvider>
+                <SearchProvider>
+                  <DefaultSeo />
+                  <Component {...pageProps} />
+                </SearchProvider>
+              </BookmarksProvider>
+            </UserStyleContextProvider>
           </ApolloProvider>
         </AuthProvider>
       </ErrorBoundary>